feat(agent): autostart the agent via launchd, not the GUI
Move launch-at-login from the GUI to the agent: a hand-written LaunchAgent plist (org.openlogi.agent, RunAtLoad, KeepAlive=true so the daemon survives crashes, no --minimized) pointing at the agent's own executable. The agent reconciles it on startup and on every IPC reload_config, and removes the legacy org.openlogi.openlogi GUI plist so the old --minimized GUI stops self-launching. The GUI's launch-at-login toggle now just saves config + sends ReloadConfig (no new IPC command); its platform/launch_agent.rs is deleted.

Hand-written plist is the unsigned/dev path; SMAppService (System Settings Login Items) is the production upgrade and needs a signed+bundled build — TODO(signing). Agent build + launch_agent plist test + workspace clippy/fmt green.
