+++ crates/openlogi-gui/src/platform/mod.rs
@@ -1,6 +1,5 @@
 //! Platform and OS integration helpers.
 
-pub mod launch_agent;
 pub mod permissions;
 pub mod single_instance;
 #[cfg(target_os = "macos")]
